requirements.txt
gradio
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
requests
